feat(accurateMostVisitedArea): add output advanced json file
