+++ .gitignore
@@ -43,3 +43,4 @@ bin/
 ### Mac OS ###
 .DS_Store
 /output.json
+/output_advanced.json
