+++ RouteAnalyzer/src/main/kotlin/model/RouteAnalysisAdvanced.kt
@@ -1,0 +1,19 @@
+package it.polito.wa2.g20.routeanalyzer.model
+
+import kotlinx.serialization.Serializable
+
+
+@Serializable
+data class Point(val latitude: Double, val longitude: Double)
+
+@Serializable
+data class AccurateMostFrequentedArea(
+    val center: Point,
+    val entriesCount: Int,
+    val areaRadiusKm: Double,
+)
+
+@Serializable
+data class RouteAnalysisAdvanced(
+    val accurateMostFrequentedArea: AccurateMostFrequentedArea?,
+)
